requirements.txt
python-pptx
Pillow
pytesseract
google-generativeai
tqdm
configparser
colorama